Thu, Oct 20, 2022, 9:07 PM +00:00

+++ scores.c
@@ -7,7 +7,7 @@ int main(void)
     int score2 = get_int("Score for second: ");
     int score3 = get_int("Score for third: ");
 
-    float n = score1 + score2 + score3;
+    int n = score1 + score2 + score3;
 
-    printf("Average %f\n", (score1 + score2 + score3) / n);
+    printf("Average %f\n", (score1 + score2 + score3) / (float) n);
 }
